+++ packages/restapi/src/lib/space/index.ts
@@ -10,5 +10,5 @@ export * from './addListeners';
 export * from './removeListeners';
 export * from './approve';
 export * from './requests';
-
+export {spaceFeed as space} from './spaceFeed';
 export * from './Space'