+++ packages/uiweb/src/lib/icons/Minimize.tsx
@@ -1,29 +1,11 @@
 import React from 'react';
 
-export const CloseSvg = ({ stroke, height, width }: { stroke?: string, height?: string, width?: string }) => {
+export const MinimizeIcon = () => {
   return (
-    <svg
-      width={width || "18.6"}
-      height={height || "19"}
-      viewBox="0 0 19 19"
-      fill='none'
-      xmlns="http://www.w3.org/2000/svg"
-      cursor="pointer"
-    >
-      <path
-        d="M14.1906 4.15625L3.97363 14.8438"
-        stroke={stroke || "#657795"}
-        stroke-width="2"
-        stroke-linecap="round"
-        stroke-linejoin="round"
-      />
-      <path
-        d="M14.1906 14.8438L3.97363 4.15625"
-        stroke={stroke || "#657795"}
-        stroke-width="2"
-        stroke-linecap="round"
-        stroke-linejoin="round"
-      />
+    <svg width="12" height="14" viewBox="0 0 12 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M6 14L0 8L1.4 6.6L6 11.175L10.6 6.6L12 8L6 14ZM6 8L0 2L1.4 0.6L6 5.175L10.6 0.6L12 2L6 8Z" fill="#62626A"/>
     </svg>
+    
+    
   );
 };