+++ packages/uiweb/src/lib/types/index.ts
@@ -1,6 +1,8 @@
-import React, { ReactElement } from 'react';
-import { ENV } from '../config';
-import { ethers } from 'ethers';
+import type { ReactElement } from 'react';
+import React from 'react';
+import type { ENV } from '../config';
+import type { ethers } from 'ethers';
+import type { ParsedResponseType, IFeeds, } from '@pushprotocol/restapi';
 
 export interface IMessageIPFS {
   fromCAIP10: string;
@@ -40,3 +42,52 @@ export type SignerType = ethers.Signer & {
   _signTypedData?: (domain: any, types: any, value: any) => Promise<string>;
   privateKey?: string;
 };
+
+export type ParsedNotificationType = ParsedResponseType & {
+  channel:string;
+};
+
+export type NotificationFeedsType = { [key: string]:ParsedNotificationType};
+export type ChatFeedsType = { [key: string]:IFeeds};
+export interface Web3NameListType {
+  [key: string]: string;
+}
+export const PUSH_TABS = {
+  CHATS: 'CHATS',
+  APP_NOTIFICATIONS: 'APP_NOTIFICATIONS'
+} as const;
+
+export const CHAT_SOCKET_TYPE = {
+  CHAT: 'chat',
+  NOTIFICATION: 'notification'
+} as const;
+
+export const PUSH_SUB_TABS = {
+  REQUESTS: 'REQUESTS',
+  SPAM:'SPAM'
+} as const;
+
+export const LOCAL_STORAGE_KEYS = {
+  CHATS: 'CHATS',
+} as const;
+
+export const SIDEBAR_PLACEHOLDER_KEYS = {
+  CHAT: 'CHAT',
+  SEARCH: 'SEARCH',
+  NOTIFICATION: 'NOTIFICATION',
+  NEW_CHAT: 'NEW_CHAT'
+} as const;
+
+export type SidebarPlaceholderKeys = (typeof SIDEBAR_PLACEHOLDER_KEYS)[keyof typeof SIDEBAR_PLACEHOLDER_KEYS];
+
+export type LocalStorageKeys = (typeof LOCAL_STORAGE_KEYS)[keyof typeof LOCAL_STORAGE_KEYS];
+export type PushTabs = (typeof PUSH_TABS)[keyof typeof PUSH_TABS];
+export type PushSubTabs = (typeof PUSH_SUB_TABS)[keyof typeof PUSH_SUB_TABS];
+export type ChatSocketType = (typeof CHAT_SOCKET_TYPE)[keyof typeof CHAT_SOCKET_TYPE];
+
+export interface FileMessageContent {
+  content: string
+  name: string
+  type: string
+  size: number
+}